+++ src/CircularString.php
@@ -144,7 +144,7 @@ class CircularString extends Curve
     #[Override]
     public function getBoundingBox() : BoundingBox
     {
-        $boundingBox = new BoundingBox();
+        $boundingBox = BoundingBox::new();
 
         foreach ($this->points as $point) {
             $boundingBox = $boundingBox->extendedWithPoint($point);