+++ src/CurvePolygon.php
@@ -137,7 +137,7 @@ class CurvePolygon extends Surface
     #[Override]
     public function getBoundingBox() : BoundingBox
     {
-        $boundingBox = new BoundingBox();
+        $boundingBox = BoundingBox::new();
 
         foreach ($this->rings as $ring) {
             $boundingBox = $boundingBox->extendedWithBoundingBox($ring->getBoundingBox());