+++ src/LineString.php
@@ -186,7 +186,7 @@ class LineString extends Curve
     #[Override]
     public function getBoundingBox() : BoundingBox
     {
-        $boundingBox = new BoundingBox();
+        $boundingBox = BoundingBox::new();
 
         foreach ($this->points as $point) {
             $boundingBox = $boundingBox->extendedWithPoint($point);