+++ src/Point.php
@@ -227,7 +227,7 @@ class Point extends Geometry
     #[Override]
     public function getBoundingBox() : BoundingBox
     {
-        return (new BoundingBox())->extendedWithPoint($this);
+        return BoundingBox::new()->extendedWithPoint($this);
     }
 
     /**