+++ src/Polygon.php
@@ -169,7 +169,7 @@ class Polygon extends Surface
     #[Override]
     public function getBoundingBox() : BoundingBox
     {
-        $boundingBox = new BoundingBox();
+        $boundingBox = BoundingBox::new();
 
         foreach ($this->rings as $ring) {
             $boundingBox = $boundingBox->extendedWithBoundingBox($ring->getBoundingBox());